Start reading commit objects
All I'm doing here is reading the commit objects in a very naive way,
but I've established I can break up the content into the correct chunks.
However, I still need to parse the dates in the content, as well as
package up the different parts into a data structure.