Use "less" as the pager for "log" command
This is based on a lot of trial and error, and finally using the
approach used by pager-rs (https://gitlab.com/imp/pager-rs).

That crate is almost perfect, but it didn't actually allow me to
configure exactly how the pager is chosen (I can only force the use of a
particular environment variable). Without this ability, I wouldn't be
able to implement the fallbacks used by git. I haven't actually
implemented said fallbacks, but that's coming in a subsequent commit.