Start parsing configuration files
This version doesn't handle all the possible syntax in git's config file
format, but it's a start. How much more I need to handle at this time
will depend on how the various system config files require me to handle.

There is one problem here, which is that the resulting config map is not
ordered. It won't matter for actual operation, but that's inconsistent
with the behavior of `git config --list`, which lists out the config
variables in the order they're declared.