List all supported commands in README
I'll be adding more commands in the near future, and many of them will
be plumbing commands that most git users aren't familiar with. Instead
of trying to provide examples for each one in the "Quick start" section,
I'll provide examples for a subset of them (mostly the porcelain
commands). I'll list out all the commands exhaustively in the new
"Supported commands" section.

As part of this documentation effort, I've added some descriptions to
the usage messages of all the commands.