Implement basic "git branch" command
Only supports listing branches, not modifying them in any way. However,
I can list all the branches and identify the current one.

In implementing this, I realized that I'm assuming the "HEAD" file will
always be a symbolic ref, which is not the case when in a "detached
HEAD" state. In that state, the "HEAD" file points directly to a commit.

This affects not only the new command, but also "git log", so I need to
think about how to resolve "HEAD" correctly in a unified way.