Refactor: return hash set of untracked files
Instead of an ordered vector. This allows us to enforce the order we
care about after the fact, when displaying the files, which means the
output of "git ls-files --others" finally matches that of the official
git implementation line for line.

The data type of the index entries is preserved because there *is* a
natural order for the entries: the order in which they appear in the
index.